refactor: separate actions from slices to fix circular import issues. fix: when joining an existing channel, use timestamp/ttl provided via qrcode Reduce number of flow-issues
