+++ BrightID/src/components/NewConnectionsScreens/pendingConnectionSlice.js
@@ -4,20 +4,11 @@ import {
   createEntityAdapter,
   createAsyncThunk,
 } from '@reduxjs/toolkit';
-import { decryptData } from '@/utils/cryptoHelper';
 import {
   removeChannel,
   selectChannelById,
 } from '@/components/NewConnectionsScreens/channelSlice';
-import { obtainKeys } from '@/utils/keychain';
-import { hash, strToUint8Array, uInt8ArrayToB64 } from '@/utils/encoding';
-import nacl from 'tweetnacl';
-import api from '@/Api/BrightId';
-import { postConnectionRequest } from '@/utils/profile';
-import { addConnection, addOperation } from '@/actions';
-import { saveImage } from '@/utils/filesystem';
-import { backupPhoto, backupUser } from '@/components/Recovery/helpers';
-import { TIME_FUDGE } from '@/utils/constants';
+import { decryptData } from '@/utils/cryptoHelper';
 
 const pendingConnectionsAdapter = createEntityAdapter();
 
@@ -35,10 +26,12 @@ const pendingConnectionsAdapter = createEntityAdapter();
    - 'signedMessage': optional - First part of signed connection message, in case user initiated connection.
    - 'timestamp': optional - Timestamp when signed message was created
  */
+
 export const pendingConnection_states = {
   INITIAL: 'INITIAL',
   DOWNLOADING: 'DOWNLOADING',
   UNCONFIRMED: 'UNCONFIRMED',
+  CONFIRMING: 'CONFIRMING',
   REJECTED: 'REJECTED',
   CONFIRMED: 'CONFIRMED',
   ERROR: 'ERROR',
@@ -50,10 +43,7 @@ export const newPendingConnection = createAsyncThunk(
     console.log(
       `new pending connection ${profileId} with signedMessage "${signedMessage}" in channel ${channelId}`,
     );
-    // TODO: using the premade "selectChannelById" selector leads to circular reference error at import, so getting
-    //  the channel manually for now
-    //  const channel = selectChannelById(getState(), channelId);
-    const channel = getState().channels.entities[channelId];
+    const channel = selectChannelById(getState(), channelId);
     // download profile
     const url = `http://${channel.ipAddress}/profile/download/${channelId}/${profileId}`;
     console.log(
@@ -76,113 +66,6 @@ export const newPendingConnection = createAsyncThunk(
       return decryptedObj;
     } else {
       throw new Error(`Missing data in profile from url: ${url}`);
-    }
-  },
-);
-
-export const confirmPendingConnectionThunk = createAsyncThunk(
-  'pendingConnections/confirmPendingConnectionThunk',
-  async (id, { getState, dispatch }) => {
-    const connection: PendingConnection = selectPendingConnectionById(
-      getState(),
-      id,
-    );
-    const channel = selectChannelById(getState(), connection.channelId);
-    console.log(`confirming connection ${id} in channel ${channel.id}`);
-
-    const {
-      user: { backupCompleted },
-    } = getState();
-
-    // validate profile state
-    if (connection.state !== pendingConnection_states.UNCONFIRMED) {
-      console.log(`Can't confirm - Connection is in state ${connection.state}`);
-      return;
-    }
-
-    let { username, secretKey } = await obtainKeys();
-    let connectionTimestamp = Date.now();
-
-    if (connection.signedMessage) {
-      // I'm responding
-      // The other user signed a connection request; we have enough info to
-      // make an API call to create the connection.
-      if (connection.timestamp > connectionTimestamp + TIME_FUDGE) {
-        throw new Error("timestamp can't be in the future");
-      }
-
-      const message = `Add Connection${connection.brightId}${username}${connection.timestamp}`;
-      const signedMessage = uInt8ArrayToB64(
-        nacl.sign.detached(strToUint8Array(message), secretKey),
-      );
-      await api.createConnection(
-        connection.brightId,
-        connection.signedMessage,
-        username,
-        signedMessage,
-        connection.timestamp,
-      );
-    } else {
-      // I'm initiating
-      const message = `Add Connection${username}${connection.brightId}${connectionTimestamp}`;
-      const signedMessage = uInt8ArrayToB64(
-        nacl.sign.detached(strToUint8Array(message), secretKey),
-      );
-
-      // bring signedmessage to peer so he knows i want to connect with him...
-      console.log(
-        `Uploading signed connection string for ${connection.name} to channel ${connection.id}`,
-      );
-      const data = {
-        signedMessage,
-        connectionTimestamp,
-      };
-      await postConnectionRequest(
-        data,
-        channel.ipAddress,
-        connection.id,
-        channel.myProfileId,
-      );
-
-      // Listen for add connection operation to be completed by other party
-      let opName = 'Add Connection';
-      let opMessage =
-        opName + username + connection.brightId + connectionTimestamp;
-      console.log(
-        `Responder opMessage: ${opMessage} - hash: ${hash(opMessage)}`,
-      );
-      const op = {
-        _key: hash(opMessage),
-        name: opName,
-        connectionTimestamp,
-      };
-      dispatch(addOperation(op));
-    }
-
-    // save connection photo
-    const filename = await saveImage({
-      imageName: connection.brightId,
-      base64Image: connection.photo,
-    });
-
-    // transfer connection from pendingConnections to established connections
-    const connectionData = {
-      id: connection.brightId,
-      name: connection.name,
-      score: connection.score,
-      // secretKey: connectUserData.secretKey,
-      // aesKey: connectUserData.aesKey,
-      connectionDate: connectionTimestamp,
-      photo: { filename },
-      status: 'initiated',
-    };
-
-    dispatch(addConnection(connectionData));
-    dispatch(confirmPendingConnection(connection.id));
-
-    if (backupCompleted) {
-      await backupUser();
-      await backupPhoto(connection.brightId, filename);
     }
   },
 );
